+++ examples/reaction-role-manager/basic.js
@@ -1,10 +1,11 @@
 const { ReactionRoleManager } = require('discord.js-collector')
 const { Client } = require("discord.js");
 const client = new Client();
+
 const reactionRoleManager = new ReactionRoleManager(client, {
-    store: true, // Enable reaction role store in a Json file
-    refreshOnBoot: true, // When the bot logs in to discord, it will give/take users who have reacted/removed the reaction while bot was offline
-    path: __dirname + '/roles.json' // Where will save the roles if store is enabled
+    storage: true, // Enable reaction role store in a Json file
+    path: __dirname + '/roles.json', // Where will save the roles if store is enabled
+    mongoDbLink: 'url mongoose link' // See here to see how setup mongoose: https://github.com/IDjinn/Discord.js-Collector/tree/dev/examples/reaction-role-manager/Note.md
 });
 
 client.on("ready", () => {
@@ -21,32 +22,36 @@ reactionRoleManager.on('reactionRoleRemove', (member, role) => {
     console.log(member.displayName + ' lose the role' + role.name)
 });
 
+// When someone removed all reactions from message
+reactionRoleManager.on('allReactionsRemove', (message) => {
+    console.log(`All reactions from message ${message.id} was removed, all roles was taken and reactions roles deleted.`)
+});
+
 client.on("message", async (message) => {
     const client = message.client;
-    const args = message.content.split(' ');
-    args.shift(); // Ignore >createReactionRole
+    const args = message.content.split(' ').slice(1);
     // Example
     // >createReactionRole @role :emoji: MessageId
     if (message.content.startsWith('>createReactionRole')) {
         const role = message.mentions.roles.first();
-        if(!role)
-            return message.reply('You need mention a role').then(m => m.delete({timeout: 1_000}));
+        if (!role)
+            return message.reply('You need mention a role').then(m => m.delete({ timeout: 1_000 }));
 
-        const emoji = client.emojis.resolveIdentifier(args.slice(1)[0])
-        if(!emoji)
-            return message.reply('You need use a valid emoji.').then(m => m.delete({timeout: 1_000}));
+        const emoji = client.emojis.resolveIdentifier(args[1])
+        if (!emoji)
+            return message.reply('You need use a valid emoji.').then(m => m.delete({ timeout: 1_000 }));
 
-        const msg = await message.channel.messages.fetch(args.slice(2)[0] || message.id);
-        if(!role)
-            return message.reply('Message not found! Wtf...').then(m => m.delete({timeout: 1_000}));
+        const msg = await message.channel.messages.fetch(args[2] || message.id);
+        if (!role)
+            return message.reply('Message not found! Wtf...').then(m => m.delete({ timeout: 1_000 }));
 
         reactionRoleManager.addRole({
             message: msg,
             role,
             emoji
         });
-        message.reply('Done').then(m => m.delete({timeout: 500}));
-        message.delete()
+        message.reply('Done').then(m => m.delete({ timeout: 500 }));
+        message.delete();
     }
 });