+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 
 ---
+# CHANGELOG - V1.4.4
+
+• Fix max counter in reaction role manager, before this version this number was inc and dec, when reaction add/remove.
+• Fix bots can win roles while reaction role manager boot
+• Improve debug logs in reaction role manager
+• Added support to mongodb (lib mongoose) in reaction role manager, as optional dependency
+
+---
+
+---
 # CHANGELOG - V1.4.3
 
 • Fix unexpected token in Constants.js